requirements.txt
django==5.0.6
prettyconf==2.2.1
psycopg2-binary==2.9.9
django-browser-reload==1.12.1
pytest==8.2.1
django-rq==2.10.2
toml==0.10.2
pytest-dependency==0.6.0
requests==2.32.3
django-admin-sortable2==2.2.6
